java-data-structures-and-algorithm/src/main/java/com/rafaelshayashi/stack/CustomStack.java
package com.rafaelshayashi.stack;

import com.rafaelshayashi.util.ArrayBaseStructure;

public class CustomStack<T> extends ArrayBaseStructure<T> {

    public static final int DEFAULT_SIZE = 10;

    public CustomStack() {
        this(DEFAULT_SIZE);
    }

    public CustomStack(int size) {
        super(size);
    }

    public void push(T element) {

        if (size == elements.length) {
            increaseSize();
        }
        elements[size] = element;
        size++;
    }

    public T pop() {

        if (isEmpty()) {
            return null;
        }

        return elements[--size];
    }

    public T peek() {
        return elements[size - 1];
    }

    @Override
    public String toString() {
        StringBuilder builder = new StringBuilder();
        builder.append("[");

        for (int i = 0; i < size - 1; i++) {
            builder.append(elements[i]).append(", ");
        }

        if (size > 0) {
            builder.append(elements[size - 1]);
        }

        builder.append("]");
        return builder.toString();
    }
}
